Initial implementation of regex path filtering on MongoDB download.

+++ oak-run-commons/src/test/java/org/apache/jackrabbit/oak/index/indexer/document/flatfile/pipelined/PipelinedIT.java
@@ -32,6 +32,7 @@ import org.apache.jackrabbit.oak.plugins.document.util.MongoConnection;
 import org.apache.jackrabbit.oak.spi.blob.MemoryBlobStore;
 import org.apache.jackrabbit.oak.spi.commit.CommitInfo;
 import org.apache.jackrabbit.oak.spi.commit.EmptyHook;
+import org.apache.jackrabbit.oak.spi.filter.PathFilter;
 import org.apache.jackrabbit.oak.spi.state.NodeBuilder;
 import org.apache.jackrabbit.oak.spi.state.NodeStore;
 import org.jetbrains.annotations.NotNull;
@@ -49,9 +50,12 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Set;
 import java.util.function.Predicate;
 
+import static org.apache.jackrabbit.oak.index.indexer.document.flatfile.pipelined.PipelinedMongoDownloadTask.OAK_INDEXER_PIPELINED_MONGO_REGEX_PATH_FILTERING;
+import static org.apache.jackrabbit.oak.index.indexer.document.flatfile.pipelined.PipelinedMongoDownloadTask.OAK_INDEXER_PIPELINED_RETRY_ON_CONNECTION_ERRORS;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
@@ -71,6 +75,8 @@ public class PipelinedIT {
         Assume.assumeTrue(MongoUtils.isAvailable());
     }
 
+    private static final PathFilter contentDamPathFilter = new PathFilter(List.of("/content/dam"), List.of());
+
     @Before
     public void setup() throws IOException {
     }
@@ -78,40 +84,76 @@ public class PipelinedIT {
     @After
     public void tear() {
         MongoConnection c = connectionFactory.getConnection();
-        c.getDatabase().drop();
+        if (c != null) {
+            c.getDatabase().drop();
+        }
     }
 
     @Test
-    public void createFFSWithPipelinedStrategy() throws Exception {
+    public void createFFS_retryOnMongoFailures_noMongoFiltering() throws Exception {
+        System.setProperty(OAK_INDEXER_PIPELINED_RETRY_ON_CONNECTION_ERRORS, "true");
+        System.setProperty(OAK_INDEXER_PIPELINED_MONGO_REGEX_PATH_FILTERING, "false");
+
+        Predicate<String> pathPredicate = s -> contentDamPathFilter.filter(s) != PathFilter.Result.EXCLUDE;
+        List<PathFilter> pathFilters = null;
+
+        testSuccessfulDownload(pathPredicate, pathFilters);
+    }
+
+    @Test
+    public void createFFS_retryOnMongoFailures_mongoFiltering() throws Exception {
+        System.setProperty(OAK_INDEXER_PIPELINED_RETRY_ON_CONNECTION_ERRORS, "true");
+        System.setProperty(OAK_INDEXER_PIPELINED_MONGO_REGEX_PATH_FILTERING, "true");
+
+        Predicate<String> pathPredicate = s -> true;
+        List<PathFilter> pathFilters = List.of(contentDamPathFilter);
+
+        testSuccessfulDownload(pathPredicate, pathFilters);
+    }
+
+    @Test
+    public void createFFS_noRetryOnMongoFailures_mongoFiltering() throws Exception {
+        System.setProperty(OAK_INDEXER_PIPELINED_RETRY_ON_CONNECTION_ERRORS, "false");
+        System.setProperty(OAK_INDEXER_PIPELINED_MONGO_REGEX_PATH_FILTERING, "true");
+
+        Predicate<String> pathPredicate = s -> true;
+        List<PathFilter> pathFilters = List.of(contentDamPathFilter);
+
+        testSuccessfulDownload(pathPredicate, pathFilters);
+    }
+
+    @Test
+    public void createFFS_noRetryOnMongoFailures_noMongoFiltering() throws Exception {
+        System.setProperty(OAK_INDEXER_PIPELINED_RETRY_ON_CONNECTION_ERRORS, "false");
+        System.setProperty(OAK_INDEXER_PIPELINED_MONGO_REGEX_PATH_FILTERING, "false");
+
+        Predicate<String> pathPredicate = s -> contentDamPathFilter.filter(s) != PathFilter.Result.EXCLUDE;
+        List<PathFilter> pathFilters = null;
+
+        testSuccessfulDownload(pathPredicate, pathFilters);
+    }
+
+    private void testSuccessfulDownload(Predicate<String> pathPredicate, List<PathFilter> pathFilters) throws CommitFailedException, IOException {
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> rwStore = createNodeStore(false);
         createContent(rwStore.right);
 
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> roStore = createNodeStore(true);
-        Predicate<String> pathPredicate = s -> s.startsWith("/content/dam");
-        PipelinedStrategy pipelinedStrategy = createStrategy(roStore, pathPredicate);
+
+        PipelinedStrategy pipelinedStrategy = createStrategy(roStore, pathPredicate, pathFilters);
 
         File file = pipelinedStrategy.createSortedStoreFile();
         assertTrue(file.exists());
-        assertEquals(Arrays.asList(new String[] {"/content/dam|{}",
-                        "/content/dam/1000|{}",
-                        "/content/dam/1000/12|{\"p1\":\"v100012\"}",
-                        "/content/dam/2022|{}",
-                        "/content/dam/2022/02|{\"p1\":\"v202202\"}",
-                        "/content/dam/2023|{\"p2\":\"v2023\"}",
-                        "/content/dam/2023/01|{\"p1\":\"v202301\"}",
-                        "/content/dam/2023/02|{}",
-                        "/content/dam/2023/02/28|{\"p1\":\"v20230228\"}"}),
-                Files.readAllLines(file.toPath()));
+        assertEquals(EXPECTED_FFS, Files.readAllLines(file.toPath()));
     }
 
     @Test
-    public void createFFSWithPipelinedStrategy_pathPredicateDoesNotMatch() throws Exception {
+    public void createFFS_pathPredicateDoesNotMatch() throws Exception {
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> rwStore = createNodeStore(false);
         createContent(rwStore.right);
 
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> roStore = createNodeStore(true);
         Predicate<String> pathPredicate = s -> s.startsWith("/content/dam/does-not-exist");
-        PipelinedStrategy pipelinedStrategy = createStrategy(roStore, pathPredicate);
+        PipelinedStrategy pipelinedStrategy = createStrategy(roStore, pathPredicate, null);
 
         File file = pipelinedStrategy.createSortedStoreFile();
 
@@ -120,7 +162,7 @@ public class PipelinedIT {
     }
 
     @Test
-    public void createFFSWithPipelinedStrategy_badNumberOfTransformThreads() throws CommitFailedException {
+    public void createFFS_badNumberOfTransformThreads() throws CommitFailedException {
         System.setProperty(PipelinedStrategy.OAK_INDEXER_PIPELINED_TRANSFORM_THREADS, "0");
 
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> rwStore = createNodeStore(false);
@@ -136,7 +178,7 @@ public class PipelinedIT {
     }
 
     @Test
-    public void createFFSWithPipelinedStrategy_badWorkingMemorySetting() throws CommitFailedException {
+    public void createFFS_badWorkingMemorySetting() throws CommitFailedException {
         System.setProperty(PipelinedStrategy.OAK_INDEXER_PIPELINED_WORKING_MEMORY_MB, "-1");
 
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> rwStore = createNodeStore(false);
@@ -161,16 +203,16 @@ public class PipelinedIT {
 
         ImmutablePair<MongoDocumentStore, DocumentNodeStore> roStore = createNodeStore(true);
         Predicate<String> pathPredicate = s -> s.startsWith("/content/dam");
-        PipelinedStrategy pipelinedStrategy = createStrategy(roStore, pathPredicate);
+        PipelinedStrategy pipelinedStrategy = createStrategy(roStore, pathPredicate, null);
 
         pipelinedStrategy.createSortedStoreFile();
     }
 
     private PipelinedStrategy createStrategy(ImmutablePair<MongoDocumentStore, DocumentNodeStore> roStore) {
-        return createStrategy(roStore, s -> true);
+        return createStrategy(roStore, s -> true, null);
     }
 
-    private PipelinedStrategy createStrategy(ImmutablePair<MongoDocumentStore, DocumentNodeStore> roStore, Predicate<String> pathPredicate) {
+    private PipelinedStrategy createStrategy(ImmutablePair<MongoDocumentStore, DocumentNodeStore> roStore, Predicate<String> pathPredicate, List<PathFilter> pathFilters) {
         DocumentNodeStore readOnlyNodeStore = roStore.right;
         MongoDocumentStore readOnlyMongoDocStore = roStore.left;
 
@@ -184,8 +226,8 @@ public class PipelinedIT {
                 new MemoryBlobStore(),
                 sortFolder.getRoot(),
                 Compression.NONE,
-                pathPredicate
-        );
+                pathPredicate,
+                pathFilters);
     }
 
     private void createContent(NodeStore rwNodeStore) throws CommitFailedException {
@@ -197,8 +239,33 @@ public class PipelinedIT {
         contentDamBuilder.child("1000").child("12").setProperty("p1", "v100012");
         contentDamBuilder.child("2023").setProperty("p2", "v2023");
         contentDamBuilder.child("2023").child("02").child("28").setProperty("p1", "v20230228");
+
+        // Other subtrees, to exercise filtering
+        rootBuilder.child("jcr:system").child("jcr:versionStorage")
+                .child("42").child("41").child("1.0").child("jcr:frozenNode").child("nodes").child("node0");
+        rootBuilder.child("home").child("users").child("system").child("cq:services").child("internal")
+                .child("dam").child("foobar").child("rep:principalPolicy").child("entry2")
+                .child("rep:restrictions");
+        rootBuilder.child("etc").child("scaffolding").child("jcr:content").child("cq:dialog")
+                .child("content").child("items").child("tabs").child("items").child("basic")
+                .child("items");
+
         rwNodeStore.merge(rootBuilder, EmptyHook.INSTANCE, CommitInfo.EMPTY);
     }
+
+    private static final List<String> EXPECTED_FFS = Arrays.asList(
+            "/|{}",
+            "/content|{}",
+            "/content/dam|{}",
+            "/content/dam/1000|{}",
+            "/content/dam/1000/12|{\"p1\":\"v100012\"}",
+            "/content/dam/2022|{}",
+            "/content/dam/2022/02|{\"p1\":\"v202202\"}",
+            "/content/dam/2023|{\"p2\":\"v2023\"}",
+            "/content/dam/2023/01|{\"p1\":\"v202301\"}",
+            "/content/dam/2023/02|{}",
+            "/content/dam/2023/02/28|{\"p1\":\"v20230228\"}"
+    );
 
     private ImmutablePair<MongoDocumentStore, DocumentNodeStore> createNodeStore(boolean readOnly) {
         MongoConnection c = connectionFactory.getConnection();
